Revert "Merge branch 'main' of https://github.com/AwesomeGecko/Project-and-Portfolio-2-Prototype"
This reverts commit d87b42ea21e0ae0079bcc9f4c17c172d9975f08a, reversing
changes made to 2d83e4aaecf21b0536912c518d7fd792021aac6f.

+++ PaP2 Prototype/Assets/Scripts/Enemy Scripts/EnemyBullet.cs
@@ -6,7 +6,7 @@ public class EnemyBullet : MonoBehaviour
 {
     [SerializeField] Rigidbody rb;
 
-    private float damage;
+    private int damage;
     private int destroyTime;
     private int speed;
     public ParticleSystem sparkParticles;
@@ -18,7 +18,7 @@ public class EnemyBullet : MonoBehaviour
         rb.velocity = (gameManager.instance.player.transform.position + ShootOffset - rb.transform.position) * speed;        
         Destroy(gameObject, destroyTime);
     }
-    public void SetBulletProperties(float damage, int destroyTime, int speed)
+    public void SetBulletProperties(int damage, int destroyTime, int speed)
     {
         this.damage = damage;
         this.destroyTime = destroyTime;
@@ -44,7 +44,7 @@ public class EnemyBullet : MonoBehaviour
 
             if (dmg != null)
             {
-                dmg.takeDamage((int)damage);
+                dmg.takeDamage(damage);
             }
 
         }
